RoutePolicyStatementMatchCollector: incorporate the BooleanExprMatchCollector
The separation of the Statement walker and the BooleanExpr walker broke because
BooleanExpr can call policies, which have Statements. That meant that any
overrides in the StatementMatchCollector were lost when going through a call or
similar.

Unifying them actually leads to simpler code, and can still be used to just walk
BooleanExpr or RoutingPolicy.

Also, since they collect from Sets e.g., Community, drop the "Match" part.

commit-id:47a81962
