+++ projects/minesweeper/src/main/java/org/batfish/minesweeper/ConfigAtomicPredicates.java
@@ -17,13 +17,12 @@ import javax.annotation.ParametersAreNonnullByDefault;
 import org.batfish.datamodel.Configuration;
 import org.batfish.datamodel.bgp.community.StandardCommunity;
 import org.batfish.datamodel.routing_policy.RoutingPolicy;
-import org.batfish.datamodel.routing_policy.expr.BooleanExprVisitor;
 import org.batfish.datamodel.routing_policy.statement.Statement;
-import org.batfish.minesweeper.aspath.BooleanExprAsPathCollector;
-import org.batfish.minesweeper.aspath.RoutePolicyStatementMatchCollector;
+import org.batfish.minesweeper.aspath.AsPathRegexCollector;
+import org.batfish.minesweeper.aspath.RoutingPolicyCollector;
 import org.batfish.minesweeper.communities.RoutePolicyStatementVarCollector;
-import org.batfish.minesweeper.env.BooleanExprSourceVrfCollector;
-import org.batfish.minesweeper.env.BooleanExprTrackCollector;
+import org.batfish.minesweeper.env.SourceVrfCollector;
+import org.batfish.minesweeper.env.TrackCollector;
 import org.batfish.minesweeper.utils.Tuple;
 
 /**
@@ -193,43 +192,31 @@ public final class ConfigAtomicPredicates {
 
   /**
    * Identifies all items of a given kind in the match expressions of the given routing policies. A
-   * {@link BooleanExprVisitor} is provided that is specific to the particular items being searched
-   * for. An optional set of additional items is also included, which is used to support
+   * {@link RoutingPolicyCollector} is provided that is specific to the particular items being
+   * searched for. An optional set of additional items is also included, which is used to support
    * user-specified constraints for symbolic analysis.
    */
   private static <T> Set<T> findAllMatchItems(
       Set<T> items,
       Collection<RoutingPolicy> policies,
       Configuration configuration,
-      BooleanExprVisitor<Set<T>, Tuple<Set<String>, Configuration>> booleanExprVisitor) {
+      RoutingPolicyCollector<T> collector) {
     ImmutableSet.Builder<T> builder = ImmutableSet.builder();
-
-    policies.forEach(
-        pol -> builder.addAll(findAllMatchItems(pol, configuration, booleanExprVisitor)));
-    builder.addAll(items.stream().collect(ImmutableSet.toImmutableSet()));
+    policies.forEach(pol -> builder.addAll(findAllMatchItems(pol, configuration, collector)));
+    builder.addAll(items);
     return builder.build();
   }
 
   /**
    * Identifies all items of a given kind in the match expressions of the given routing policy. A
-   * {@link BooleanExprVisitor} is provided that is specific to the particular items being searched
-   * for. An optional set of additional items is also included, which is used to support
+   * {@link RoutingPolicyCollector} is provided that is specific to the particular items being
+   * searched for. An optional set of additional items is also included, which is used to support
    * user-specified constraints for symbolic analysis.
    */
   private static <T> Set<T> findAllMatchItems(
-      RoutingPolicy policy,
-      Configuration configuration,
-      BooleanExprVisitor<Set<T>, Tuple<Set<String>, Configuration>> booleanExprVisitor) {
-    Set<T> items = new HashSet<>();
-    List<Statement> stmts = policy.getStatements();
-    stmts.forEach(
-        stmt ->
-            items.addAll(
-                stmt.accept(
-                    new RoutePolicyStatementMatchCollector<>(booleanExprVisitor),
-                    new Tuple<>(
-                        new HashSet<>(Collections.singleton(policy.getName())), configuration))));
-    return items;
+      RoutingPolicy policy, Configuration configuration, RoutingPolicyCollector<T> collector) {
+    Set<String> visited = new HashSet<>(Collections.singleton(policy.getName()));
+    return collector.visitAll(policy.getStatements(), new Tuple<>(visited, configuration));
   }
 
   /**
@@ -241,8 +228,7 @@ public final class ConfigAtomicPredicates {
       Set<SymbolicAsPathRegex> asPathRegexes,
       Collection<RoutingPolicy> policies,
       Configuration configuration) {
-    return findAllMatchItems(
-        asPathRegexes, policies, configuration, new BooleanExprAsPathCollector());
+    return findAllMatchItems(asPathRegexes, policies, configuration, new AsPathRegexCollector());
   }
 
   /**
@@ -267,8 +253,7 @@ public final class ConfigAtomicPredicates {
    */
   private static Set<String> findAllTracks(
       Collection<RoutingPolicy> policies, Configuration configuration) {
-    return findAllMatchItems(
-        ImmutableSet.of(), policies, configuration, new BooleanExprTrackCollector());
+    return findAllMatchItems(ImmutableSet.of(), policies, configuration, new TrackCollector());
   }
 
   /**
@@ -280,8 +265,7 @@ public final class ConfigAtomicPredicates {
    */
   private static Set<String> findAllSourceVrfs(
       Collection<RoutingPolicy> policies, Configuration configuration) {
-    return findAllMatchItems(
-        ImmutableSet.of(), policies, configuration, new BooleanExprSourceVrfCollector());
+    return findAllMatchItems(ImmutableSet.of(), policies, configuration, new SourceVrfCollector());
   }
 
   public RegexAtomicPredicates<CommunityVar> getStandardCommunityAtomicPredicates() {
