+++ projects/minesweeper/src/main/java/org/batfish/minesweeper/NextHopInterfaceCollector.java
@@ -6,12 +6,12 @@ import javax.annotation.ParametersAreNonnullByDefault;
 import org.batfish.datamodel.Configuration;
 import org.batfish.datamodel.routing_policy.expr.BooleanExpr;
 import org.batfish.datamodel.routing_policy.expr.MatchInterface;
-import org.batfish.minesweeper.aspath.BooleanExprMatchCollector;
+import org.batfish.minesweeper.aspath.RoutingPolicyCollector;
 import org.batfish.minesweeper.utils.Tuple;
 
 /** Collect all next-hop interface names (see {@link MatchInterface}) in a {@link BooleanExpr}. */
 @ParametersAreNonnullByDefault
-public class NextHopInterfaceCollector extends BooleanExprMatchCollector<String> {
+public class NextHopInterfaceCollector extends RoutingPolicyCollector<String> {
   @Override
   public Set<String> visitMatchInterface(
       MatchInterface matchInterface, Tuple<Set<String>, Configuration> arg) {
