+++ projects/minesweeper/src/main/java/org/batfish/minesweeper/collectors/CalledPolicyCollector.java
@@ -3,9 +3,10 @@ package org.batfish.minesweeper.collectors;
 import com.google.common.collect.ImmutableSet;
 import java.util.Set;
 import org.batfish.datamodel.Configuration;
+import org.batfish.datamodel.routing_policy.expr.CallExpr;
 import org.batfish.datamodel.routing_policy.statement.CallStatement;
 import org.batfish.datamodel.routing_policy.statement.Statement;
-import org.batfish.minesweeper.aspath.RoutePolicyStatementMatchCollector;
+import org.batfish.minesweeper.aspath.RoutingPolicyCollector;
 import org.batfish.minesweeper.utils.Tuple;
 
 /**
@@ -13,11 +14,11 @@ import org.batfish.minesweeper.utils.Tuple;
  * only returns the calls made by the original policy, excluding any other policy calls made by the
  * callees.
  */
-public final class RoutePolicyStatementCallCollector
-    extends RoutePolicyStatementMatchCollector<String> {
+public final class CalledPolicyCollector extends RoutingPolicyCollector<String> {
 
-  public RoutePolicyStatementCallCollector() {
-    super(new RoutePolicyBooleanExprCallCollector());
+  @Override
+  public Set<String> visitCallExpr(CallExpr callExpr, Tuple<Set<String>, Configuration> arg) {
+    return ImmutableSet.of(callExpr.getCalledPolicyName());
   }
 
   @Override
