+++ projects/minesweeper/src/main/java/org/batfish/minesweeper/collectors/PrefixListNameCollector.java
@@ -4,17 +4,17 @@ import com.google.common.collect.ImmutableSet;
 import java.util.Set;
 import javax.annotation.ParametersAreNonnullByDefault;
 import org.batfish.datamodel.Configuration;
-import org.batfish.datamodel.routing_policy.expr.BooleanExpr;
 import org.batfish.datamodel.routing_policy.expr.HasRoute;
 import org.batfish.datamodel.routing_policy.expr.MatchPrefixSet;
 import org.batfish.datamodel.routing_policy.expr.NamedPrefixSet;
-import org.batfish.minesweeper.aspath.BooleanExprMatchCollector;
+import org.batfish.minesweeper.aspath.RoutingPolicyCollector;
 import org.batfish.minesweeper.utils.Tuple;
 
-/** Collect all prefix-list names in a {@link BooleanExpr}. */
+/**
+ * Collect all prefix-list names in a {@link org.batfish.datamodel.routing_policy.RoutingPolicy}.
+ */
 @ParametersAreNonnullByDefault
-public class RouteFilterBooleanExprCollector extends BooleanExprMatchCollector<String> {
-
+public class PrefixListNameCollector extends RoutingPolicyCollector<String> {
   @Override
   public Set<String> visitHasRoute(HasRoute hasRoute, Tuple<Set<String>, Configuration> arg) {
     if (hasRoute.getExpr() instanceof NamedPrefixSet) {
