+++ projects/minesweeper/src/main/java/org/batfish/minesweeper/question/comparepeergrouppolicies/RoutingPolicyContextDiff.java
@@ -14,10 +14,9 @@ import java.util.TreeSet;
 import javax.annotation.Nonnull;
 import org.batfish.datamodel.Configuration;
 import org.batfish.datamodel.routing_policy.RoutingPolicy;
-import org.batfish.minesweeper.aspath.RoutePolicyStatementMatchCollector;
 import org.batfish.minesweeper.collectors.AsPathNameBooleanExprCollector;
-import org.batfish.minesweeper.collectors.RoutingPolicyStatementCommunityCollector;
-import org.batfish.minesweeper.collectors.RoutingPolicyStatementRouteFilterCollector;
+import org.batfish.minesweeper.collectors.CommunityNameCollector;
+import org.batfish.minesweeper.collectors.PrefixListNameCollector;
 import org.batfish.minesweeper.utils.Tuple;
 
 /**
@@ -96,21 +95,21 @@ public final class RoutingPolicyContextDiff implements Comparable<RoutingPolicyC
   public boolean differ(RoutingPolicy currentPolicy) {
     // Find all the references/names used by the routing policy
     Set<String> communityNames =
-        new RoutingPolicyStatementCommunityCollector()
+        new CommunityNameCollector()
             .visitAll(
                 currentPolicy.getStatements(),
                 new Tuple<>(
                     new HashSet<>(Collections.singleton(currentPolicy.getName())),
                     this._currentConfig));
     Set<String> routeFilterNames =
-        new RoutingPolicyStatementRouteFilterCollector()
+        new PrefixListNameCollector()
             .visitAll(
                 currentPolicy.getStatements(),
                 new Tuple<>(
                     new HashSet<>(Collections.singleton(currentPolicy.getName())),
                     this._currentConfig));
     Set<String> asPathNames =
-        new RoutePolicyStatementMatchCollector<>(new AsPathNameBooleanExprCollector())
+        new AsPathNameBooleanExprCollector()
             .visitAll(
                 currentPolicy.getStatements(),
                 new Tuple<>(
