+++ projects/minesweeper/src/test/java/org/batfish/minesweeper/collectors/CalledPolicyCollectorTest.java
@@ -5,8 +5,8 @@ import static org.hamcrest.Matchers.empty;
 import static org.junit.Assert.assertEquals;
 
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
-import java.io.IOException;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
@@ -14,6 +14,7 @@ import java.util.Set;
 import org.batfish.datamodel.Configuration;
 import org.batfish.datamodel.ConfigurationFormat;
 import org.batfish.datamodel.NetworkFactory;
+import org.batfish.datamodel.TraceElement;
 import org.batfish.datamodel.routing_policy.RoutingPolicy;
 import org.batfish.datamodel.routing_policy.expr.CallExpr;
 import org.batfish.datamodel.routing_policy.expr.Conjunction;
@@ -23,33 +24,89 @@ import org.batfish.datamodel.routing_policy.expr.FirstMatchChain;
 import org.batfish.datamodel.routing_policy.expr.Not;
 import org.batfish.datamodel.routing_policy.expr.TrackSucceeded;
 import org.batfish.datamodel.routing_policy.expr.WithEnvironmentExpr;
+import org.batfish.datamodel.routing_policy.statement.BufferedStatement;
 import org.batfish.datamodel.routing_policy.statement.CallStatement;
 import org.batfish.datamodel.routing_policy.statement.If;
 import org.batfish.datamodel.routing_policy.statement.Statements;
+import org.batfish.datamodel.routing_policy.statement.TraceableStatement;
 import org.batfish.minesweeper.utils.Tuple;
 import org.junit.Before;
 import org.junit.Test;
 
-public class RoutePolicyBooleanExprCallCollectorTest {
+/** Tests for {@link CalledPolicyCollector}. */
+public class CalledPolicyCollectorTest {
   private static final String HOSTNAME = "hostname";
   private static final String RM1 = "RM1";
   private static final String RM2 = "RM2";
   private static final String RM3 = "RM3";
   private static final String RM4 = "RM4";
-  private RoutePolicyBooleanExprCallCollector _collector;
+  private Configuration _baseConfig;
   private NetworkFactory _nf;
-  private Configuration _config;
+  private CalledPolicyCollector _collector;
 
   @Before
-  public void setup() throws IOException {
+  public void setup() {
     _nf = new NetworkFactory();
     Configuration.Builder cb =
         _nf.configurationBuilder()
             .setHostname(HOSTNAME)
-            .setConfigurationFormat(ConfigurationFormat.CUMULUS_CONCATENATED);
-    _config = cb.build();
+            .setConfigurationFormat(ConfigurationFormat.CISCO_IOS);
+    _baseConfig = cb.build();
+    _nf.vrfBuilder().setOwner(_baseConfig).setName(Configuration.DEFAULT_VRF_NAME).build();
 
-    _collector = new RoutePolicyBooleanExprCallCollector();
+    _collector = new CalledPolicyCollector();
+  }
+
+  @Test
+  public void testVisitBufferedStatement() {
+    BufferedStatement bs =
+        new BufferedStatement(
+            new If(new CallExpr(RM1), ImmutableList.of(Statements.ExitAccept.toStaticStatement())));
+
+    Set<String> result =
+        _collector.visitBufferedStatement(bs, new Tuple<>(new HashSet<>(), _baseConfig));
+
+    assertEquals(ImmutableSet.of(RM1), result);
+  }
+
+  @Test
+  public void testVisitIf() {
+    If ifstmt = new If(new CallExpr(RM2), ImmutableList.of(new CallStatement(RM1)));
+    Set<String> result = _collector.visitIf(ifstmt, new Tuple<>(new HashSet<>(), _baseConfig));
+
+    assertEquals(ImmutableSet.of(RM1, RM2), result);
+  }
+
+  @Test
+  public void testVisitTraceableStatement() {
+    TraceableStatement traceableStatement =
+        new TraceableStatement(
+            TraceElement.of("statement"), ImmutableList.of(new CallStatement(RM2)));
+
+    Set<String> result =
+        _collector.visitTraceableStatement(
+            traceableStatement, new Tuple<>(new HashSet<>(), _baseConfig));
+    assertEquals(ImmutableSet.of(RM2), result);
+  }
+
+  @Test
+  public void testVisitCallStatement() {
+
+    RoutingPolicy rm2 =
+        _nf.routingPolicyBuilder()
+            .setName(RM2)
+            .setOwner(_baseConfig)
+            .addStatement(new CallStatement("ignored"))
+            .build();
+
+    RoutingPolicy rm1 =
+        _nf.routingPolicyBuilder().setName(RM1).addStatement(new CallStatement(RM2)).build();
+
+    _baseConfig.setRoutingPolicies(ImmutableMap.of(RM2, rm2, RM1, rm1));
+
+    assertEquals(
+        ImmutableSet.of(RM2),
+        _collector.visitAll(rm1.getStatements(), new Tuple<>(new HashSet<>(), _baseConfig)));
   }
 
   @Test
@@ -57,7 +114,7 @@ public class RoutePolicyBooleanExprCallCollectorTest {
 
     Conjunction c = new Conjunction(ImmutableList.of(new CallExpr(RM2), new CallExpr(RM3)));
 
-    Set<String> result = _collector.visitConjunction(c, new Tuple<>(new HashSet<>(), _config));
+    Set<String> result = _collector.visitConjunction(c, new Tuple<>(new HashSet<>(), _baseConfig));
 
     assertEquals(ImmutableSet.of(RM2, RM3), result);
   }
@@ -69,7 +126,7 @@ public class RoutePolicyBooleanExprCallCollectorTest {
         new ConjunctionChain(ImmutableList.of(new CallExpr(RM2), new CallExpr(RM3)));
 
     Set<String> result =
-        _collector.visitConjunctionChain(cc, new Tuple<>(new HashSet<>(), _config));
+        _collector.visitConjunctionChain(cc, new Tuple<>(new HashSet<>(), _baseConfig));
 
     assertEquals(ImmutableSet.of(RM2, RM3), result);
   }
@@ -79,7 +136,7 @@ public class RoutePolicyBooleanExprCallCollectorTest {
 
     Disjunction d = new Disjunction(ImmutableList.of(new CallExpr(RM2), new CallExpr(RM3)));
 
-    Set<String> result = _collector.visitDisjunction(d, new Tuple<>(new HashSet<>(), _config));
+    Set<String> result = _collector.visitDisjunction(d, new Tuple<>(new HashSet<>(), _baseConfig));
 
     assertEquals(ImmutableSet.of(RM2, RM3), result);
   }
@@ -91,7 +148,7 @@ public class RoutePolicyBooleanExprCallCollectorTest {
         new FirstMatchChain(ImmutableList.of(new CallExpr(RM2), new CallExpr(RM3)));
 
     Set<String> result =
-        _collector.visitFirstMatchChain(fmc, new Tuple<>(new HashSet<>(), _config));
+        _collector.visitFirstMatchChain(fmc, new Tuple<>(new HashSet<>(), _baseConfig));
 
     assertEquals(ImmutableSet.of(RM2, RM3), result);
   }
@@ -100,7 +157,7 @@ public class RoutePolicyBooleanExprCallCollectorTest {
   public void testVisitTrackSucceeded() {
     assertThat(
         _collector.visitTrackSucceeded(
-            new TrackSucceeded("foo"), new Tuple<>(new HashSet<>(), _config)),
+            new TrackSucceeded("foo"), new Tuple<>(new HashSet<>(), _baseConfig)),
         empty());
   }
 
@@ -109,7 +166,7 @@ public class RoutePolicyBooleanExprCallCollectorTest {
 
     Not n = new Not(new CallExpr(RM1));
 
-    Set<String> result = _collector.visitNot(n, new Tuple<>(new HashSet<>(), _config));
+    Set<String> result = _collector.visitNot(n, new Tuple<>(new HashSet<>(), _baseConfig));
 
     assertEquals(ImmutableSet.of(RM1), result);
   }
@@ -124,7 +181,7 @@ public class RoutePolicyBooleanExprCallCollectorTest {
     wee.setPostTrueStatements(ImmutableList.of(new CallStatement(RM4)));
 
     Set<String> result =
-        _collector.visitWithEnvironmentExpr(wee, new Tuple<>(new HashSet<>(), _config));
+        _collector.visitWithEnvironmentExpr(wee, new Tuple<>(new HashSet<>(), _baseConfig));
 
     assertEquals(ImmutableSet.of(RM1, RM2, RM3, RM4), result);
   }
@@ -132,10 +189,10 @@ public class RoutePolicyBooleanExprCallCollectorTest {
   @Test
   public void testCallExprNested() {
     Map<String, RoutingPolicy> routingPolicies = new HashMap<>();
-    _config.setRoutingPolicies(routingPolicies);
+    _baseConfig.setRoutingPolicies(routingPolicies);
     RoutingPolicy rm1 =
         _nf.routingPolicyBuilder()
-            .setOwner(_config)
+            .setOwner(_baseConfig)
             .setName("RM1")
             .addStatement(
                 new If(
@@ -145,7 +202,7 @@ public class RoutePolicyBooleanExprCallCollectorTest {
     routingPolicies.put(RM1, rm1);
 
     Set<String> result =
-        _collector.visitCallExpr(new CallExpr(RM1), new Tuple<>(new HashSet<>(), _config));
+        _collector.visitCallExpr(new CallExpr(RM1), new Tuple<>(new HashSet<>(), _baseConfig));
 
     assertEquals(ImmutableSet.of(RM1), result);
   }
